requirements.txt
certifi==2021.5.30
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4 freeze
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==2.0.1
pickle-mixin==1.0.2
Pillow==8.0.1
requests==2.25.0
Werkzeug==1.0.1
wincertstore==0.2
